+++ src/main/java/ru/yandex/practicum/filmorate/storage/likes_film/LikesFilmDbStorage.java
@@ -11,14 +11,14 @@ public class LikesFilmDbStorage implements LikesFilmStorage {
     private final JdbcTemplate jdbcTemplate;
 
     @Override
-    public void addLikeFilm(int filmId, int userId) {
+    public void addLikeFilm(long filmId, long userId) {
         String sqlQueryForLikeMark = "INSERT INTO likes_film (film_id, user_id) VALUES (?, ?)";
 
         jdbcTemplate.update(sqlQueryForLikeMark, filmId, userId);
     }
 
     @Override
-    public void deleteLikeFilm(int filmId, int userId) {
+    public void deleteLikeFilm(long filmId, long userId) {
         String sqlQueryForUnLikeMark = "DELETE FROM likes_film WHERE film_id = ? AND user_id = ?";
 
         jdbcTemplate.update(sqlQueryForUnLikeMark, filmId, userId);
